refactor server_variables 1. move functions to separated files 2. render HTML at once 3. refactor function name to PMA_getHtmlForSubPageHeader
