+++ server_plugins.php
@@ -51,7 +51,7 @@ ksort($plugins);
 /**
  * Displays the page
  */
-$response->addHTML(PMA_getSubPageHeader('plugins'));
+$response->addHTML(PMA_getHtmlForSubPageHeader('plugins'));
 $response->addHTML(PMA_getPluginAndModuleInfo($plugins, $modules));
 
 exit;
